+++ .cargo/config
@@ -1,0 +1,3 @@
+[build]
+rustflags = ["-C", "linker=clang"]
+#rustflags = ["-C", "link-arg=-fuse-ld=lld"]
