+++ node/src/models/object.rs
