+++ node/src/rpc/file_transfer/mod.rs
@@ -10,11 +10,11 @@ use std::sync::Arc;
 use samizdat_common::cipher::TransferCipher;
 use samizdat_common::Hash;
 
-use crate::cache::{CollectionItem, ObjectRef};
 use crate::cli;
+use crate::models::{CollectionItem, ObjectRef};
 
 const MAX_HEADER_LENGTH: usize = 4_096;
-const MAX_STREAM_SIZE: usize = crate::cache::CHUNK_SIZE * 2;
+const MAX_STREAM_SIZE: usize = crate::models::CHUNK_SIZE * 2;
 
 fn read_error_to_io(error: ReadToEndError) -> io::Error {
     match error {
