+++ node/src/rpc/mod.rs
@@ -1,5 +1,6 @@
 mod connection_manager;
 mod file_transfer;
+mod node_server;
 mod reconnect;
 
 pub use reconnect::Reconnect;
@@ -19,143 +20,11 @@ use samizdat_common::quic;
 use samizdat_common::rpc::*;
 use samizdat_common::{ContentRiddle, Hash};
 
-use crate::cache::{CollectionItem, ObjectRef, SeriesRef, SeriesItem};
 use crate::cli;
+use crate::models::{ObjectRef, SeriesItem, SeriesRef};
 
 use connection_manager::{ConnectionManager, DropMode};
-
-#[derive(Clone)]
-struct NodeServer {
-    connection_manager: Arc<ConnectionManager>,
-}
-
-#[tarpc::server]
-impl Node for NodeServer {
-    async fn resolve_object(
-        self,
-        _: context::Context,
-        resolution: Arc<Resolution>,
-    ) -> ResolutionResponse {
-        log::info!("got object {:?}", resolution);
-
-        let object = match ObjectRef::find(&resolution.content_riddle) {
-            Some(object) => object,
-            None => {
-                log::info!("hash not found for resolution");
-                return ResolutionResponse::NOT_FOUND;
-            }
-        };
-
-        // Code smell?
-        let hash = object.hash;
-
-        log::info!("found hash {}", object.hash);
-        let peer_addr = match resolution.message_riddle.resolve::<SocketAddr>(&hash) {
-            Some(socket_addr) => socket_addr,
-            None => {
-                log::warn!("failed to resolve message riddle after resolving content riddle");
-                return ResolutionResponse::FOUND;
-            }
-        };
-
-        log::info!("found peer at {}", peer_addr);
-
-        tokio::spawn(
-            async move {
-                let new_connection = self
-                    .connection_manager
-                    .punch_hole_to(peer_addr, DropMode::DropIncoming)
-                    .await?;
-                file_transfer::send_object(&new_connection.connection, &object).await
-            }
-            .map(move |outcome| {
-                outcome
-                    .map_err(|err| log::error!("failed to send {} to {}: {}", hash, peer_addr, err))
-            }),
-        );
-
-        return ResolutionResponse::FOUND;
-    }
-
-    async fn resolve_item(
-        self,
-        _: context::Context,
-        resolution: Arc<Resolution>,
-    ) -> ResolutionResponse {
-        log::info!("got item {:?}", resolution);
-
-        let item = match CollectionItem::find(&resolution.content_riddle) {
-            Ok(Some(item)) => item,
-            Ok(None) => {
-                log::info!("hash not found for resolution");
-                return ResolutionResponse::NOT_FOUND;
-            }
-            Err(e) => {
-                log::error!("error looking for hash: {}", e);
-                return ResolutionResponse::NOT_FOUND;
-            }
-        };
-
-        // Code smell?
-        let hash = item.locator().hash();
-
-        log::info!("found hash {}", hash);
-        let peer_addr = match resolution.message_riddle.resolve::<SocketAddr>(&hash) {
-            Some(socket_addr) => socket_addr,
-            None => {
-                log::warn!("failed to resolve message riddle after resolving content riddle");
-                return ResolutionResponse::FOUND;
-            }
-        };
-
-        log::info!("found peer at {}", peer_addr);
-
-        tokio::spawn(
-            async move {
-                let new_connection = self
-                    .connection_manager
-                    .punch_hole_to(peer_addr, DropMode::DropIncoming)
-                    .await?;
-                file_transfer::send_item(&new_connection.connection, item).await
-            }
-            .map(move |outcome| {
-                outcome
-                    .map_err(|err| log::error!("failed to send {} to {}: {}", hash, peer_addr, err))
-            }),
-        );
-
-        return ResolutionResponse::FOUND;
-    }
-
-    async fn resolve_latest(
-        self,
-        _: context::Context,
-        latest: Arc<LatestRequest>,
-    ) -> Option<LatestResponse> {
-        if let Some(series) = SeriesRef::find(&latest.key_riddle) {
-            match series.get_latest_fresh() {
-                Ok(None) => None,
-                Ok(Some(mut latest)) => {
-                    let cipher_key = latest.public_key().hash();
-                    let rand = Hash::rand();
-                    let cipher = TransferCipher::new(&cipher_key, &rand);
-                    latest.erase_freshness();
-
-                    Some(LatestResponse {
-                        rand,
-                        series: cipher.encrypt_opaque(&latest),
-                    })
-                }
-                Err(err) => {
-                    log::warn!("{}", err);
-                    None
-                }
-            }
-        } else {
-            None
-        }
-    }
-}
+use node_server::NodeServer;
 
 pub struct HubConnectionInner {
     client: HubClient,
@@ -335,25 +204,21 @@ impl HubConnection {
         }
     }
 
-    pub async fn get_latest(
-        &self,
-        series: &SeriesRef,
-    ) -> Result<Option<SeriesItem>, crate::Error> {
+    pub async fn get_latest(&self, series: &SeriesRef) -> Result<Option<SeriesItem>, crate::Error> {
         let key_riddle = ContentRiddle::new(&series.public_key.hash());
         let inner = self.inner.get().await;
 
-        let response = inner.client.get_latest(
-            context::current(),
-            LatestRequest {
-            key_riddle,
-        }).await?;
+        let response = inner
+            .client
+            .get_latest(context::current(), LatestRequest { key_riddle })
+            .await?;
 
         let mut most_recent: Option<SeriesItem> = None;
 
         for candidate in response {
             let cipher = TransferCipher::new(&series.public_key.hash(), &candidate.rand);
             let candidate_item: SeriesItem = candidate.series.decrypt_with(&cipher)?;
-            
+
             if !candidate_item.is_valid() {
                 log::warn!("received invalid candidate item: {:?}", candidate_item);
                 continue;
